requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
reservoirpy>=0.3.5
supabase>=2.0.0
gradio>=4.0.0
huggingface_hub>=0.20.0
